+++ CHANGELOG.md
@@ -2,7 +2,21 @@
 
 All notable changes to this project will be documented in this file.
 
-This project is semver based
+This project is [semver](https://semver.org/) based
+
+## [0.2] - Model List is updated from Stable Horde API
+
+### Added
+ * Selection Model list is updated from the API and user is
+   informed when new ones are installed.
+ * User is invited to create an api_key with option to visit with
+   the browser
+ * Warnings about model usage are seen by user
+
+### Changed
+ * The settings file is readable only by the user, no others,
+   except admins
+ * Bugfixes
 
 ## [0.1] - First Release of LibreOffice Macro